added camelcasing and resource not existing condition

+++ src/main/java/com/ecommerce/ProductService/services/category/CategoryServiceImpl.java
@@ -41,6 +41,9 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Long id) {
+        if(!repo.existsById(id)){
+            throw new ResourceNotFoundException("Category not found with id: " + id);
+        }
         repo.deleteById(id);
     }
